@@ -3,11 +3,9 @@ description: A Flutter plugin to integrate the stripe libraries for iOS and Andr
 version: 1.0.9
 homepage: https://github.com/jonasbark/flutter_stripe_payment
 
-environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.0 <2.0.0"
-
 dependencies:
+  build_runner: ^1.0.0
+  json_annotation:  ^3.0.1
   flutter:
     sdk: flutter
 
@@ -15,7 +13,19 @@ flutter:
   plugin:
     platforms:
       android:
-        package: de.jonasbark.stripepayment
+        package: de.jonasbark.stripe_payment
         pluginClass: StripePaymentPlugin
       ios:
         pluginClass: StripePaymentPlugin
+    
+dev_dependencies:
+  json_serializable: ^3.3.0
+  flutter_test:
+    sdk: flutter
+  integration_test: ^1.0.1
+  flutter_driver:
+    sdk: flutter
+                
+environment:
+  sdk: ">=2.6.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5 <2.0.0"
