@@ -2,9 +2,8 @@ package com.gettipsi.stripe;
 
 import android.app.Activity;
 import android.content.Intent;
-import androidx.annotation.NonNull;
 import android.util.Log;
-
+import androidx.annotation.NonNull;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReadableMap;
 import com.gettipsi.stripe.util.ArgCheck;
@@ -13,6 +12,7 @@ import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GoogleApiAvailability;
 import com.google.android.gms.wallet.WalletConstants;
 import com.stripe.android.BuildConfig;
+import com.stripe.android.Stripe;
 
 public abstract class PayFlow {
 
@@ -26,18 +26,19 @@ public abstract class PayFlow {
     this.activityProvider = activityProvider;
   }
 
-  public static PayFlow create(Fun0<Activity> activityProvider) {
-    return new GoogleApiPayFlowImpl(activityProvider);
+  public static PayFlow create(Fun0<Activity> activityProvider, Fun0<Stripe> stripeProvider) {
+    return new GoogleApiPayFlowImpl(activityProvider, stripeProvider);
   }
 
   private static boolean isValidEnvironment(int environment) {
-    return environment == WalletConstants.ENVIRONMENT_TEST ||
-      environment == WalletConstants.ENVIRONMENT_PRODUCTION;
+    return environment == WalletConstants.ENVIRONMENT_TEST
+        || environment == WalletConstants.ENVIRONMENT_PRODUCTION;
   }
 
   private static boolean isEnvironmentChangeAttempt(int oldEnvironment, int newEnvironment) {
-    return oldEnvironment != newEnvironment && isValidEnvironment(oldEnvironment) &&
-      isValidEnvironment(newEnvironment);
+    return oldEnvironment != newEnvironment
+        && isValidEnvironment(oldEnvironment)
+        && isValidEnvironment(newEnvironment);
   }
 
   protected int getEnvironment() {
@@ -79,11 +80,17 @@ public abstract class PayFlow {
     return Errors.getDescription(getErrorCodes(), key);
   }
 
+  abstract void paymentMethodFromAndroidPay(final ReadableMap payParams, final Promise promise);
+
   abstract void paymentRequestWithAndroidPay(final ReadableMap payParams, final Promise promise);
 
-  abstract void deviceSupportsAndroidPay(boolean isExistingPaymentMethodRequired, final Promise promise);
+  abstract void deviceSupportsAndroidPay(
+      boolean isExistingPaymentMethodRequired, final Promise promise);
 
-  abstract boolean onActivityResult(Activity activity, int requestCode, int resultCode, Intent data);
+  abstract void potentiallyAvailableNativePayNetworks(final Promise promise);
+
+  abstract boolean onActivityResult(
+      Activity activity, int requestCode, int resultCode, Intent data);
 
   public static boolean isPlayServicesAvailable(@NonNull Activity activity) {
     ArgCheck.nonNull(activity);
