@@ -1,85 +1,123 @@
 package com.gettipsi.stripe;
 
+import static com.gettipsi.stripe.Errors.AUTHENTICATION_FAILED;
+import static com.gettipsi.stripe.Errors.CANCELLED;
+import static com.gettipsi.stripe.Errors.FAILED;
+import static com.gettipsi.stripe.Errors.UNEXPECTED;
+import static com.gettipsi.stripe.Errors.getDescription;
+import static com.gettipsi.stripe.Errors.getErrorCode;
+import static com.gettipsi.stripe.Errors.toErrorCode;
+import static com.gettipsi.stripe.util.Converters.convertPaymentIntentResultToWritableMap;
+import static com.gettipsi.stripe.util.Converters.convertPaymentMethodToWritableMap;
+import static com.gettipsi.stripe.util.Converters.convertSetupIntentResultToWritableMap;
+import static com.gettipsi.stripe.util.Converters.convertSourceToWritableMap;
+import static com.gettipsi.stripe.util.Converters.convertTokenToWritableMap;
+import static com.gettipsi.stripe.util.Converters.createBankAccountTokenParams;
+import static com.gettipsi.stripe.util.Converters.createCard;
+import static com.gettipsi.stripe.util.Converters.getBooleanOrNull;
+import static com.gettipsi.stripe.util.Converters.getMapOrNull;
+import static com.gettipsi.stripe.util.Converters.getStringOrNull;
+import static com.gettipsi.stripe.util.InitializationOptions.ANDROID_PAY_MODE_KEY;
+import static com.gettipsi.stripe.util.InitializationOptions.ANDROID_PAY_MODE_PRODUCTION;
+import static com.gettipsi.stripe.util.InitializationOptions.ANDROID_PAY_MODE_TEST;
+import static com.gettipsi.stripe.util.InitializationOptions.PUBLISHABLE_KEY;
+import static com.stripe.android.model.StripeIntent.Status.Canceled;
+import static com.stripe.android.model.StripeIntent.Status.RequiresAction;
+import static com.stripe.android.model.StripeIntent.Status.RequiresCapture;
+import static com.stripe.android.model.StripeIntent.Status.RequiresConfirmation;
+import static com.stripe.android.model.StripeIntent.Status.Succeeded;
+
 import android.app.Activity;
-import android.content.DialogInterface;
-import android.content.Intent;
+import android.content.Context;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.text.TextUtils;
+import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import com.facebook.react.bridge.*;
+import com.facebook.react.bridge.Promise;
+import com.facebook.react.bridge.ReactMethod;
+import com.facebook.react.bridge.ReadableMap;
 import com.gettipsi.stripe.util.ArgCheck;
 import com.gettipsi.stripe.util.Converters;
-import com.gettipsi.stripe.util.Fun0;
 import com.google.android.gms.wallet.WalletConstants;
-import com.stripe.android.*;
-import com.stripe.android.model.*;
-
-import de.jonasbark.stripepayment.StripeDialog;
-import io.flutter.app.FlutterActivity;
+import com.stripe.android.ApiResultCallback;
+import com.stripe.android.AppInfo;
+import com.stripe.android.PaymentIntentResult;
+import com.stripe.android.SetupIntentResult;
+import com.stripe.android.Stripe;
+import com.stripe.android.model.Address;
+import com.stripe.android.model.ConfirmPaymentIntentParams;
+import com.stripe.android.model.ConfirmSetupIntentParams;
+import com.stripe.android.model.PaymentMethod;
+import com.stripe.android.model.PaymentMethodCreateParams;
+import com.stripe.android.model.Source;
+import com.stripe.android.model.Source.Redirect;
+import com.stripe.android.model.SourceParams;
+import com.stripe.android.model.StripeIntent;
+import com.stripe.android.model.Token;
+import de.jonasbark.stripe_payment.ActivityRegistry;
+import de.jonasbark.stripe_payment.StripeDialog;
 import io.flutter.plugin.common.PluginRegistry;
-import kotlin.Unit;
-import kotlin.jvm.functions.Function1;
-
+import io.flutter.plugin.common.PluginRegistry.ActivityResultListener;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicReference;
+import kotlin.Unit;
+import org.jetbrains.annotations.NotNull;
 
-import static com.gettipsi.stripe.Errors.*;
-import static com.gettipsi.stripe.util.Converters.*;
-import static com.gettipsi.stripe.util.InitializationOptions.*;
-import static com.stripe.android.model.StripeIntent.Status.*;
+public class StripeModule {
 
-public class StripeModule extends ReactContextBaseJavaModule {
-
-
-  private static final String MODULE_NAME = StripeModule.class.getSimpleName();
+  private static final String TAG = "StripeModule";
 
   // If you change these, make sure to also change:
   //  ios/TPSStripe/TPSStripeManager
   // Relevant Docs:
-  // - https://stripe.dev/stripe-ios/docs/Classes/STPAppInfo.html https://stripe.dev/stripe-android/com/stripe/android/AppInfo.html
+  // - https://stripe.dev/stripe-ios/docs/Classes/STPAppInfo.html
+  // https://stripe.dev/stripe-android/com/stripe/android/AppInfo.html
   // - https://stripe.com/docs/building-plugins#setappinfo
-  private static final String APP_INFO_NAME    = "tipsi-stripe";
-  private static final String APP_INFO_URL     = "https://github.com/tipsi/tipsi-stripe";
+  private static final String APP_INFO_NAME = "tipsi-stripe";
+  private static final String APP_INFO_URL = "https://github.com/tipsi/tipsi-stripe";
   private static final String APP_INFO_VERSION = "8.x";
   public static final String CLIENT_SECRET = "clientSecret";
 
-  private static StripeModule sInstance = null;
+  @NonNull private Context applicationContext;
 
-  public static StripeModule getInstance() {
-    return sInstance;
-  }
+  @Nullable public Activity activity = null;
+
+  private final ActivityRegistry activityRegistry;
 
   public Stripe getStripe() {
     return mStripe;
   }
 
-  @Nullable
-  private Promise mCreateSourcePromise;
+  @Nullable private Promise mCreateSourcePromise;
 
-  @Nullable
-  private Source mCreatedSource;
+  @Nullable private Source mCreatedSource;
 
   private String mPublicKey;
   private Stripe mStripe;
   private PayFlow mPayFlow;
   private ReadableMap mErrorCodes;
 
-  private final PluginRegistry.ActivityResultListener mActivityEventListener = new PluginRegistry.ActivityResultListener() {
+  private final PluginRegistry.ActivityResultListener activityResultListener =
+      (requestCode, resultCode, data) -> {
+        if (activity != null) {
+          return getPayFlow().onActivityResult(activity, requestCode, resultCode, data);
+        }
+        return false;
+      };
 
-    @Override
-    public boolean onActivityResult(int requestCode, int resultCode, Intent data) {
-      boolean handled = getPayFlow().onActivityResult(activity, requestCode, resultCode, data);
-      return handled;
-    }
-  };
+  public StripeModule(
+      @NotNull Context applicationContext,
+      @Nullable Activity activity,
+      ActivityRegistry activityRegistry) {
+    this.applicationContext = applicationContext;
+    this.activity = activity;
+    this.activityRegistry = activityRegistry;
 
-  public StripeModule(PluginRegistry.Registrar registrar, Activity activity) {
-    super(activity, registrar);
-    registrar.addActivityResultListener(mActivityEventListener);
-
-    sInstance = this;
+    activityRegistry.addListener(activityResultListener);
   }
 
   @ReactMethod
@@ -94,12 +132,15 @@ public class StripeModule extends ReactContextBaseJavaModule {
 
       mPublicKey = newPubKey;
       Stripe.setAppInfo(AppInfo.create(APP_INFO_NAME, APP_INFO_VERSION, APP_INFO_URL));
-      mStripe = new Stripe(getReactApplicationContext(), mPublicKey);
+      mStripe = new Stripe(applicationContext, mPublicKey);
+
       getPayFlow().setPublishableKey(mPublicKey);
     }
 
     if (newAndroidPayMode != null) {
-      ArgCheck.isTrue(ANDROID_PAY_MODE_TEST.equals(newAndroidPayMode) || ANDROID_PAY_MODE_PRODUCTION.equals(newAndroidPayMode));
+      ArgCheck.isTrue(
+          ANDROID_PAY_MODE_TEST.equals(newAndroidPayMode)
+              || ANDROID_PAY_MODE_PRODUCTION.equals(newAndroidPayMode));
 
       getPayFlow().setEnvironment(androidPayModeToEnvironment(newAndroidPayMode));
     }
@@ -112,11 +153,7 @@ public class StripeModule extends ReactContextBaseJavaModule {
 
   private PayFlow getPayFlow() {
     if (mPayFlow == null) {
-      mPayFlow = PayFlow.create(
-        new Fun0<Activity>() { public Activity call() {
-          return getCurrentActivity();
-        }}
-      );
+      mPayFlow = PayFlow.create(() -> activity, () -> mStripe);
     }
 
     return mPayFlow;
@@ -124,7 +161,9 @@ public class StripeModule extends ReactContextBaseJavaModule {
 
   private static int androidPayModeToEnvironment(@NonNull String androidPayMode) {
     ArgCheck.notEmptyString(androidPayMode);
-    return ANDROID_PAY_MODE_TEST.equals(androidPayMode.toLowerCase()) ? WalletConstants.ENVIRONMENT_TEST : WalletConstants.ENVIRONMENT_PRODUCTION;
+    return ANDROID_PAY_MODE_TEST.equals(androidPayMode.toLowerCase())
+        ? WalletConstants.ENVIRONMENT_TEST
+        : WalletConstants.ENVIRONMENT_PRODUCTION;
   }
 
   @ReactMethod
@@ -138,12 +177,17 @@ public class StripeModule extends ReactContextBaseJavaModule {
   }
 
   @ReactMethod
+  public void potentiallyAvailableNativePayNetworks(final Promise promise) {
+    getPayFlow().potentiallyAvailableNativePayNetworks(promise);
+  }
+
+  @ReactMethod
   public void setStripeAccount(final String stripeAccount) {
     ArgCheck.notEmptyString(mPublicKey);
     if (stripeAccount == null) {
-      mStripe = new Stripe(getReactApplicationContext(), mPublicKey);
+      mStripe = new Stripe(applicationContext, mPublicKey);
     } else {
-      mStripe = new Stripe(getReactApplicationContext(), mPublicKey, stripeAccount);
+      mStripe = new Stripe(applicationContext, mPublicKey, stripeAccount);
     }
   }
 
@@ -153,18 +197,21 @@ public class StripeModule extends ReactContextBaseJavaModule {
       ArgCheck.nonNull(mStripe);
       ArgCheck.notEmptyString(mPublicKey);
 
-      mStripe.createToken(
-        createCard(cardData),
-        mPublicKey,
-        new TokenCallback() {
-          public void onSuccess(Token token) {
-            promise.resolve(convertTokenToWritableMap(token));
-          }
-          public void onError(Exception error) {
-            error.printStackTrace();
-            promise.reject(toErrorCode(error), error.getMessage());
-          }
-        });
+      mStripe.createCardToken(
+          createCard(cardData),
+          null,
+          new ApiResultCallback<Token>() {
+            @Override
+            public void onSuccess(Token token) {
+              promise.resolve(convertTokenToWritableMap(token));
+            }
+
+            @Override
+            public void onError(@NotNull Exception error) {
+              error.printStackTrace();
+              promise.reject(toErrorCode(error), error.getMessage());
+            }
+          });
     } catch (Exception e) {
       promise.reject(toErrorCode(e), e.getMessage());
     }
@@ -177,18 +224,18 @@ public class StripeModule extends ReactContextBaseJavaModule {
       ArgCheck.notEmptyString(mPublicKey);
 
       mStripe.createBankAccountToken(
-        createBankAccount(accountData),
-        mPublicKey,
-        null,
-        new TokenCallback() {
-          public void onSuccess(Token token) {
-            promise.resolve(convertTokenToWritableMap(token));
-          }
-          public void onError(Exception error) {
-            error.printStackTrace();
-            promise.reject(toErrorCode(error), error.getMessage());
-          }
-        });
+          createBankAccountTokenParams(accountData),
+          null,
+          new ApiResultCallback<Token>() {
+            public void onSuccess(Token token) {
+              promise.resolve(convertTokenToWritableMap(token));
+            }
+
+            public void onError(@NonNull Exception error) {
+              error.printStackTrace();
+              promise.reject(toErrorCode(error), error.getMessage());
+            }
+          });
     } catch (Exception e) {
       promise.reject(toErrorCode(e), e.getMessage());
     }
@@ -196,31 +243,27 @@ public class StripeModule extends ReactContextBaseJavaModule {
 
   @ReactMethod
   public void paymentRequestWithCardForm(ReadableMap params, final Promise promise) {
-    Activity currentActivity = getCurrentActivity();
+    Activity currentActivity = activity;
     try {
       ArgCheck.nonNull(currentActivity);
       ArgCheck.notEmptyString(mPublicKey);
-      final StripeDialog cardDialog = StripeDialog.newInstance(
-        ""
-      );
+      final StripeDialog cardDialog = StripeDialog.newInstance("");
       cardDialog.setStripeInstance(mStripe);
-      cardDialog.setTokenListener(new Function1<PaymentMethod, Unit>() {
-        @Override
-        public Unit invoke(PaymentMethod paymentMethod) {
-          promise.resolve(Converters.convertPaymentMethodToWritableMap(paymentMethod));
-          return Unit.INSTANCE;
-        }
-      });
-      cardDialog.setOnCancelListener(new DialogInterface.OnCancelListener() {
-        @Override
-        public void onCancel(DialogInterface dialog) {
-          promise.reject(CANCELLED, CANCELLED);
-        }
-      });
+      cardDialog.setTokenListener(
+          paymentMethod -> {
+            promise.resolve(Converters.convertPaymentMethodToWritableMap(paymentMethod));
+            return Unit.INSTANCE;
+          });
+      cardDialog.setOnCancelListener(dialog -> promise.reject(CANCELLED, CANCELLED));
       cardDialog.show(currentActivity.getFragmentManager(), "AddNewCard");
     } catch (Exception e) {
       promise.reject(toErrorCode(e), e.getMessage());
     }
+  }
+
+  @ReactMethod
+  public void paymentMethodFromAndroidPay(final ReadableMap payParams, final Promise promise) {
+    getPayFlow().paymentMethodFromAndroidPay(payParams, promise);
   }
 
   @ReactMethod
@@ -229,107 +272,105 @@ public class StripeModule extends ReactContextBaseJavaModule {
   }
 
   private void attachPaymentResultActivityListener(final Promise promise) {
-    ActivityEventListener ael = new BaseActivityEventListener() {
+    final ActivityResultListener ael =
+        (requestCode, resultCode, data) ->
+            mStripe.onPaymentResult(
+                requestCode,
+                data,
+                new ApiResultCallback<PaymentIntentResult>() {
+                  @Override
+                  public void onSuccess(@NonNull PaymentIntentResult result) {
+                    StripeIntent.Status resultingStatus = result.getIntent().getStatus();
 
-      @Override
-      public void onActivityResult(Activity a, int requestCode, int resultCode, Intent data) {
-        final ActivityEventListener ael = this;
+                    if (Succeeded.equals(resultingStatus)
+                        || RequiresCapture.equals(resultingStatus)
+                        || RequiresConfirmation.equals(resultingStatus)) {
+                      promise.resolve(convertPaymentIntentResultToWritableMap(result));
+                    } else {
+                      if (Canceled.equals(resultingStatus)
+                          || RequiresAction.equals(resultingStatus)) {
+                        promise.reject(CANCELLED, CANCELLED); // TODO - normalize the message
+                      } else {
+                        promise.reject(FAILED, FAILED);
+                      }
+                    }
+                  }
 
-        mStripe.onPaymentResult(requestCode, data, new ApiResultCallback<PaymentIntentResult>() {
-          @Override
-          public void onSuccess(@NonNull PaymentIntentResult result) {
-            removeActivityEventListener(ael);
+                  @Override
+                  public void onError(@NonNull Exception e) {
+                    e.printStackTrace();
+                    promise.reject(toErrorCode(e), e.getMessage());
+                  }
+                });
 
-            StripeIntent.Status resultingStatus = result.getIntent().getStatus();
+    activityRegistry.addListener(ael);
 
-            if (Succeeded.equals(resultingStatus) ||
-                RequiresCapture.equals(resultingStatus) ||
-                RequiresConfirmation.equals(resultingStatus)) {
-              promise.resolve(convertPaymentIntentResultToWritableMap(result));
-            } else {
-              if (Canceled.equals(resultingStatus) ||
-                  RequiresAction.equals(resultingStatus)
-              ) {
-                promise.reject(CANCELLED, CANCELLED);      // TODO - normalize the message
-              } else {
-                promise.reject(FAILED, FAILED);
-              }
-            }
-          }
-
-          @Override
-          public void onError(@NonNull Exception e) {
-            removeActivityEventListener(ael);
-            e.printStackTrace();
-            promise.reject(toErrorCode(e), e.getMessage());
+    promise.setWhenComplete(
+        () -> {
+          if (activityRegistry.removeListener(ael)) {
+            hasSetupResultListener.set(false);
           }
         });
-      }
-
-      @Override
-      public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        onActivityResult(null, requestCode, resultCode, data);
-      }
-    };
-    addActivityEventListener(ael);
   }
 
+  private AtomicBoolean hasSetupResultListener = new AtomicBoolean(false);
+
   private void attachSetupResultActivityListener(final Promise promise) {
-    ActivityEventListener ael = new BaseActivityEventListener() {
-      @Override
-      public void onActivityResult(Activity a, int requestCode, int resultCode, Intent data) {
-        final ActivityEventListener ael = this;
+    ActivityResultListener ael =
+        (requestCode, resultCode, data) ->
+            mStripe.onSetupResult(
+                requestCode,
+                data,
+                new ApiResultCallback<SetupIntentResult>() {
+                  @Override
+                  public void onSuccess(@NonNull SetupIntentResult result) {
+                    try {
+                      switch (result.getIntent().getStatus()) {
+                        case Canceled:
+                          // The Setup Intent was canceled, so reject the promise with a predefined
+                          // code.
+                          promise.reject(CANCELLED, "The SetupIntent was canceled by the user.");
+                          break;
+                        case RequiresAction:
+                        case RequiresPaymentMethod:
+                          promise.reject(AUTHENTICATION_FAILED, "The user failed authentication.");
+                          break;
+                        case Succeeded:
+                          promise.resolve(convertSetupIntentResultToWritableMap(result));
+                          break;
+                        case RequiresCapture:
+                        case RequiresConfirmation:
+                        default:
+                          promise.reject(UNEXPECTED, "Unexpected state");
+                      }
+                    } catch (Exception e) {
+                      Log.e(TAG, "Unexpected error", e);
+                      promise.reject(UNEXPECTED, "Unexpected error");
+                    }
+                  }
 
-        mStripe.onSetupResult(requestCode, data, new ApiResultCallback<SetupIntentResult>() {
-          @Override
-          public void onSuccess(@NonNull SetupIntentResult result) {
-            removeActivityEventListener(ael);
+                  @Override
+                  public void onError(@NonNull Exception e) {
+                    Errors.raiseFlutterError(e, promise);
+                  }
+                });
 
-            try {
-              switch (result.getIntent().getStatus()) {
-                case Canceled:
-                  // The Setup Intent was canceled, so reject the promise with a predefined code.
-                  promise.reject(CANCELLED, "The SetupIntent was canceled by the user.");
-                  break;
-                case RequiresAction:
-                case RequiresPaymentMethod:
-                  promise.reject(AUTHENTICATION_FAILED, "The user failed authentication.");
-                  break;
-                case Succeeded:
-                  promise.resolve(convertSetupIntentResultToWritableMap(result));
-                  break;
-                case RequiresCapture:
-                case RequiresConfirmation:
-                default:
-                  promise.reject(UNEXPECTED, "Unexpected state");
-              }
-            } catch (Exception e) {
-              promise.reject(UNEXPECTED, "Unexpected error");
-            }
-          }
+    if (!hasSetupResultListener.getAndSet(true)) {
+      activityRegistry.addListener(ael);
+    }
 
-          @Override
-          public void onError(@NonNull Exception e) {
-            removeActivityEventListener(ael);
-            e.printStackTrace();
-            promise.reject(toErrorCode(e), e.getMessage());
+    promise.setWhenComplete(
+        () -> {
+          if (activityRegistry.removeListener(ael)) {
+            hasSetupResultListener.set(false);
           }
         });
-      }
-
-      @Override
-      public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        onActivityResult(null, requestCode, resultCode, data);
-      }
-    };
-    addActivityEventListener(ael);
   }
 
   @ReactMethod
   public void confirmPaymentIntent(final ReadableMap options, final Promise promise) {
     attachPaymentResultActivityListener(promise);
 
-    Activity activity = getCurrentActivity();
     if (activity != null) {
       mStripe.confirmPayment(activity, extractConfirmPaymentIntentParams(options));
     }
@@ -340,7 +381,6 @@ public class StripeModule extends ReactContextBaseJavaModule {
     attachPaymentResultActivityListener(promise);
 
     String clientSecret = options.getString(CLIENT_SECRET);
-    Activity activity = getCurrentActivity();
     if (activity != null) {
       mStripe.authenticatePayment(activity, clientSecret);
     }
@@ -350,7 +390,6 @@ public class StripeModule extends ReactContextBaseJavaModule {
   public void confirmSetupIntent(final ReadableMap options, final Promise promise) {
     attachSetupResultActivityListener(promise);
 
-    Activity activity = getCurrentActivity();
     if (activity != null) {
       mStripe.confirmSetupIntent(activity, extractConfirmSetupIntentParams(options));
     }
@@ -361,69 +400,69 @@ public class StripeModule extends ReactContextBaseJavaModule {
     attachSetupResultActivityListener(promise);
 
     String clientSecret = options.getString(CLIENT_SECRET);
-    Activity activity = getCurrentActivity();
     if (activity != null) {
       mStripe.authenticateSetup(activity, clientSecret);
     }
   }
-
 
   @ReactMethod
   public void createPaymentMethod(final ReadableMap options, final Promise promise) {
 
     PaymentMethodCreateParams pmcp = extractPaymentMethodCreateParams(options);
 
-    mStripe.createPaymentMethod(pmcp, new ApiResultCallback<PaymentMethod>() {
+    mStripe.createPaymentMethod(
+        pmcp,
+        new ApiResultCallback<PaymentMethod>() {
 
-      @Override
-      public void onError(Exception error) {
-        promise.reject(toErrorCode(error), error.getMessage());
-      }
+          @Override
+          public void onError(@NonNull Exception error) {
+            promise.reject(toErrorCode(error), error.getMessage());
+          }
 
-      @Override
-      public void onSuccess(PaymentMethod paymentMethod) {
-        promise.resolve(convertPaymentMethodToWritableMap(paymentMethod));
-      }
-    });
+          @Override
+          public void onSuccess(PaymentMethod paymentMethod) {
+            promise.resolve(convertPaymentMethodToWritableMap(paymentMethod));
+          }
+        });
   }
-
 
   @ReactMethod
   public void createSourceWithParams(final ReadableMap options, final Promise promise) {
-
     SourceParams sourceParams = extractSourceParams(options);
 
     ArgCheck.nonNull(sourceParams);
 
-    mStripe.createSource(sourceParams, new SourceCallback() {
-      @Override
-      public void onError(Exception error) {
-        promise.reject(toErrorCode(error));
-      }
-
-      @Override
-      public void onSuccess(Source source) {
-        if (Source.SourceFlow.REDIRECT.equals(source.getFlow())) {
-          Activity currentActivity = getCurrentActivity();
-          if (currentActivity == null) {
-            promise.reject(
-              getErrorCode(mErrorCodes, "activityUnavailable"),
-              getDescription(mErrorCodes, "activityUnavailable")
-            );
-          } else {
-            mCreateSourcePromise = promise;
-            mCreatedSource = source;
-            String redirectUrl = source.getRedirect().getUrl();
-            Intent browserIntent = new Intent(currentActivity, OpenBrowserActivity.class)
-                .addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_SINGLE_TOP)
-                .putExtra(OpenBrowserActivity.EXTRA_URL, redirectUrl);
-            currentActivity.startActivity(browserIntent);
+    mStripe.createSource(
+        sourceParams,
+        new ApiResultCallback<Source>() {
+          @Override
+          public void onError(@NonNull Exception error) {
+            promise.reject(toErrorCode(error));
           }
-        } else {
-          promise.resolve(convertSourceToWritableMap(source));
-        }
-      }
-    });
+
+          @Override
+          public void onSuccess(Source source) {
+            if (source.getRedirect() != null) {
+              if (activity == null) {
+                promise.reject(
+                    getErrorCode(mErrorCodes, "activityUnavailable"),
+                    getDescription(mErrorCodes, "activityUnavailable"));
+              } else {
+                mCreateSourcePromise = promise;
+                mCreatedSource = source;
+                //            String redirectUrl = source.getRedirect().getUrl();
+                //            Intent browserIntent = new Intent(activity, OpenBrowserActivity.class)
+                //                .addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP |
+                // Intent.FLAG_ACTIVITY_SINGLE_TOP)
+                //                .putExtra(OpenBrowserActivity.EXTRA_URL, redirectUrl);
+                //            activity.startActivity(browserIntent);
+                // TODO
+              }
+            } else {
+              promise.resolve(convertSourceToWritableMap(source));
+            }
+          }
+        });
   }
 
   private ConfirmSetupIntentParams extractConfirmSetupIntentParams(final ReadableMap options) {
@@ -432,13 +471,14 @@ public class StripeModule extends ReactContextBaseJavaModule {
     String returnURL = getStringOrNull(options, "returnURL");
     String clientSecret = options.getString("clientSecret");
     ConfirmSetupIntentParams csip = null;
-    if (returnURL ==  null) {
+    if (returnURL == null) {
       returnURL = "stripejs://use_stripe_sdk/return_url";
     }
 
     if (paymentMethod != null) {
-      csip = ConfirmSetupIntentParams.create(extractPaymentMethodCreateParams(paymentMethod),
-        clientSecret, returnURL);
+      csip =
+          ConfirmSetupIntentParams.create(
+              extractPaymentMethodCreateParams(paymentMethod), clientSecret, returnURL);
     } else if (paymentMethodId != null) {
       csip = ConfirmSetupIntentParams.create(paymentMethodId, clientSecret, returnURL);
     }
@@ -453,13 +493,13 @@ public class StripeModule extends ReactContextBaseJavaModule {
     String clientSecret = options.getString("clientSecret");
 
     ReadableMap paymentMethod = getMapOrNull(options, "paymentMethod");
-    String paymentMethodId = getStringOrNull(options,"paymentMethodId");
+    String paymentMethodId = getStringOrNull(options, "paymentMethodId");
 
     // ReadableMap source = options.getMap("source");
     // String sourceId = getStringOrNull(options,"sourceId");
 
     String returnURL = getStringOrNull(options, "returnURL");
-    if (returnURL ==  null) {
+    if (returnURL == null) {
       returnURL = "stripejs://use_stripe_sdk/return_url";
     }
     boolean savePaymentMethod = getBooleanOrNull(options, "savePaymentMethod", false);
@@ -471,41 +511,45 @@ public class StripeModule extends ReactContextBaseJavaModule {
     if (paymentMethod != null) {
 
       PaymentMethodCreateParams pmcp = extractPaymentMethodCreateParams(paymentMethod);
-      cpip = ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(pmcp, clientSecret, returnURL, savePaymentMethod, extraParams);
+      cpip =
+          ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(
+              pmcp, clientSecret, returnURL, savePaymentMethod, extraParams);
 
-    // Create with Payment Method ID
+      // Create with Payment Method ID
     } else if (paymentMethodId != null) {
 
-      cpip = ConfirmPaymentIntentParams.createWithPaymentMethodId(paymentMethodId, clientSecret, returnURL, savePaymentMethod, extraParams);
+      cpip =
+          ConfirmPaymentIntentParams.createWithPaymentMethodId(
+              paymentMethodId, clientSecret, returnURL, savePaymentMethod, extraParams);
 
-    // Create with Source
-    /**
-    Support for creating a Source while confirming a PaymentIntent is not being included
-    at this time, however, for compatibility with existing saved Sources, you can still confirm
-    a payment intent using a pre-existing Source by specifying its 'sourceId', as shown in the next
-    branch
-    */
-    /*
-    } else if (source != null) {
-      SourceParams sourceParams = extractSourceParams(source);
-      cpip = ConfirmPaymentIntentParams.createWithSourceParams(sourceParams, clientSecret, returnURL, savePaymentMethod, extraParams);
-    */
+      // Create with Source
+      /**
+       * Support for creating a Source while confirming a PaymentIntent is not being included at
+       * this time, however, for compatibility with existing saved Sources, you can still confirm a
+       * payment intent using a pre-existing Source by specifying its 'sourceId', as shown in the
+       * next branch
+       */
+      /*
+      } else if (source != null) {
+        SourceParams sourceParams = extractSourceParams(source);
+        cpip = ConfirmPaymentIntentParams.createWithSourceParams(sourceParams, clientSecret, returnURL, savePaymentMethod, extraParams);
+      */
 
-    // Create with Source ID
-    /**
-    If you have a sourceId, pass it into the paymentMethodId parameter instead!
-    The payment_method parameter of a payment intent is fully compatible with Sources.
-    Reference: https://stripe.com/docs/api/payment_intents/confirm#confirm_payment_intent-payment_method
-    */
-    /*
-    } else if (sourceId != null) {
-      cpip = ConfirmPaymentIntentParams.createWithSourceId(sourceId, clientSecret, returnURL, savePaymentMethod, extraParams);
-    */
+      // Create with Source ID
+      /**
+       * If you have a sourceId, pass it into the paymentMethodId parameter instead! The
+       * payment_method parameter of a payment intent is fully compatible with Sources. Reference:
+       * https://stripe.com/docs/api/payment_intents/confirm#confirm_payment_intent-payment_method
+       */
+      /*
+      } else if (sourceId != null) {
+        cpip = ConfirmPaymentIntentParams.createWithSourceId(sourceId, clientSecret, returnURL, savePaymentMethod, extraParams);
+      */
 
-    /**
-    This branch can be used if the client secret refers to a payment intent that already
-    has payment method information and just needs to be confirmed.
-    */
+      /**
+       * This branch can be used if the client secret refers to a payment intent that already has
+       * payment method information and just needs to be confirmed.
+       */
     } else {
       cpip = ConfirmPaymentIntentParams.create(clientSecret, returnURL);
     }
@@ -537,22 +581,24 @@ public class StripeModule extends ReactContextBaseJavaModule {
       ReadableMap addressParams = getMapOrNull(options, "address");
 
       if (addressParams != null) {
-        address = new Address.Builder().
-          setCity(getStringOrNull(addressParams, "city")).
-          setCountry(addressParams.getString("country")).
-          setLine1(getStringOrNull(addressParams, "line1")).
-          setLine2(getStringOrNull(addressParams, "line2")).
-          setPostalCode(getStringOrNull(addressParams, "postalCode")).
-          setState(getStringOrNull(addressParams, "state")).
-          build();
+        address =
+            new Address.Builder()
+                .setCity(getStringOrNull(addressParams, "city"))
+                .setCountry(addressParams.getString("country"))
+                .setLine1(getStringOrNull(addressParams, "line1"))
+                .setLine2(getStringOrNull(addressParams, "line2"))
+                .setPostalCode(getStringOrNull(addressParams, "postalCode"))
+                .setState(getStringOrNull(addressParams, "state"))
+                .build();
       }
 
-      billingDetails = new PaymentMethod.BillingDetails.Builder().
-        setAddress(address).
-        setEmail(getStringOrNull(billingDetailsParams, "email")).
-        setName(getStringOrNull(billingDetailsParams,"name")).
-        setPhone(getStringOrNull(billingDetailsParams,"phone")).
-        build();
+      billingDetails =
+          new PaymentMethod.BillingDetails.Builder()
+              .setAddress(address)
+              .setEmail(getStringOrNull(billingDetailsParams, "email"))
+              .setName(getStringOrNull(billingDetailsParams, "name"))
+              .setPhone(getStringOrNull(billingDetailsParams, "phone"))
+              .build();
     }
 
     if (cardParams != null) {
@@ -560,20 +606,17 @@ public class StripeModule extends ReactContextBaseJavaModule {
       if (token != null) {
         card = PaymentMethodCreateParams.Card.create(token);
       } else {
-        card = new PaymentMethodCreateParams.Card.Builder().
-          setCvc(cardParams.getString("cvc")).
-          setExpiryMonth(cardParams.getInt("expMonth")).
-          setExpiryYear(cardParams.getInt("expYear")).
-          setNumber(cardParams.getString("number")).
-          build();
+        card =
+            new PaymentMethodCreateParams.Card.Builder()
+                .setCvc(cardParams.getString("cvc"))
+                .setExpiryMonth(cardParams.getInt("expMonth"))
+                .setExpiryYear(cardParams.getInt("expYear"))
+                .setNumber(cardParams.getString("number"))
+                .build();
       }
     }
 
-    return PaymentMethodCreateParams.create(
-      card,
-      billingDetails,
-      metadata
-    );
+    return PaymentMethodCreateParams.create(card, billingDetails, metadata);
   }
 
   private SourceParams extractSourceParams(final ReadableMap options) {
@@ -581,58 +624,65 @@ public class StripeModule extends ReactContextBaseJavaModule {
     SourceParams sourceParams = null;
     switch (sourceType) {
       case "alipay":
-        sourceParams = SourceParams.createAlipaySingleUseParams(
-          options.getInt("amount"),
-          options.getString("currency"),
-          getStringOrNull(options, "name"),
-          getStringOrNull(options, "email"),
-          options.getString("returnURL"));
+        sourceParams =
+            SourceParams.createAlipaySingleUseParams(
+                options.getInt("amount"),
+                options.getString("currency"),
+                getStringOrNull(options, "name"),
+                getStringOrNull(options, "email"),
+                options.getString("returnURL"));
         break;
       case "bancontact":
-        sourceParams = SourceParams.createBancontactParams(
-          options.getInt("amount"),
-          options.getString("name"),
-          options.getString("returnURL"),
-          getStringOrNull(options, "statementDescriptor"),
-          options.getString("preferredLanguage"));
+        sourceParams =
+            SourceParams.createBancontactParams(
+                options.getInt("amount"),
+                options.getString("name"),
+                options.getString("returnURL"),
+                getStringOrNull(options, "statementDescriptor"),
+                options.getString("preferredLanguage"));
         break;
       case "giropay":
-        sourceParams = SourceParams.createGiropayParams(
-          options.getInt("amount"),
-          options.getString("name"),
-          options.getString("returnURL"),
-          getStringOrNull(options, "statementDescriptor"));
+        sourceParams =
+            SourceParams.createGiropayParams(
+                options.getInt("amount"),
+                options.getString("name"),
+                options.getString("returnURL"),
+                getStringOrNull(options, "statementDescriptor"));
         break;
       case "ideal":
-        sourceParams = SourceParams.createIdealParams(
-          options.getInt("amount"),
-          options.getString("name"),
-          options.getString("returnURL"),
-          getStringOrNull(options, "statementDescriptor"),
-          getStringOrNull(options, "bank"));
+        sourceParams =
+            SourceParams.createIdealParams(
+                options.getInt("amount"),
+                options.getString("name"),
+                options.getString("returnURL"),
+                getStringOrNull(options, "statementDescriptor"),
+                getStringOrNull(options, "bank"));
         break;
       case "sepaDebit":
-        sourceParams = SourceParams.createSepaDebitParams(
-          options.getString("name"),
-          options.getString("iban"),
-          getStringOrNull(options, "addressLine1"),
-          options.getString("city"),
-          options.getString("postalCode"),
-          options.getString("country"));
+        sourceParams =
+            SourceParams.createSepaDebitParams(
+                options.getString("name"),
+                options.getString("iban"),
+                getStringOrNull(options, "addressLine1"),
+                options.getString("city"),
+                options.getString("postalCode"),
+                options.getString("country"));
         break;
       case "sofort":
-        sourceParams = SourceParams.createSofortParams(
-          options.getInt("amount"),
-          options.getString("returnURL"),
-          options.getString("country"),
-          getStringOrNull(options, "statementDescriptor"));
+        sourceParams =
+            SourceParams.createSofortParams(
+                options.getInt("amount"),
+                options.getString("returnURL"),
+                options.getString("country"),
+                getStringOrNull(options, "statementDescriptor"));
         break;
       case "threeDSecure":
-        sourceParams = SourceParams.createThreeDSecureParams(
-          options.getInt("amount"),
-          options.getString("currency"),
-          options.getString("returnURL"),
-          options.getString("card"));
+        sourceParams =
+            SourceParams.createThreeDSecureParams(
+                options.getInt("amount"),
+                options.getString("currency"),
+                options.getString("returnURL"),
+                options.getString("card"));
         break;
       case "card":
         sourceParams = SourceParams.createCardParams(Converters.createCard(options));
@@ -641,7 +691,6 @@ public class StripeModule extends ReactContextBaseJavaModule {
     return sourceParams;
   }
 
-
   void processRedirect(@Nullable Uri redirectData) {
     if (mCreatedSource == null || mCreateSourcePromise == null) {
 
@@ -649,11 +698,9 @@ public class StripeModule extends ReactContextBaseJavaModule {
     }
 
     if (redirectData == null) {
-
       mCreateSourcePromise.reject(
-        getErrorCode(mErrorCodes, "redirectCancelled"),
-        getDescription(mErrorCodes, "redirectCancelled")
-      );
+          getErrorCode(mErrorCodes, "redirectCancelled"),
+          getDescription(mErrorCodes, "redirectCancelled"));
       mCreatedSource = null;
       mCreateSourcePromise = null;
       return;
@@ -662,9 +709,8 @@ public class StripeModule extends ReactContextBaseJavaModule {
     final String clientSecret = redirectData.getQueryParameter("client_secret");
     if (!mCreatedSource.getClientSecret().equals(clientSecret)) {
       mCreateSourcePromise.reject(
-        getErrorCode(mErrorCodes, "redirectNoSource"),
-        getDescription(mErrorCodes, "redirectNoSource")
-      );
+          getErrorCode(mErrorCodes, "redirectNoSource"),
+          getDescription(mErrorCodes, "redirectNoSource"));
       mCreatedSource = null;
       mCreateSourcePromise = null;
       return;
@@ -673,9 +719,8 @@ public class StripeModule extends ReactContextBaseJavaModule {
     final String sourceId = redirectData.getQueryParameter("source");
     if (!mCreatedSource.getId().equals(sourceId)) {
       mCreateSourcePromise.reject(
-        getErrorCode(mErrorCodes, "redirectWrongSourceId"),
-        getDescription(mErrorCodes, "redirectWrongSourceId")
-      );
+          getErrorCode(mErrorCodes, "redirectWrongSourceId"),
+          getDescription(mErrorCodes, "redirectWrongSourceId"));
       mCreatedSource = null;
       mCreateSourcePromise = null;
       return;
@@ -701,27 +746,24 @@ public class StripeModule extends ReactContextBaseJavaModule {
       protected void onPostExecute(Source source) {
         if (source != null) {
           switch (source.getStatus()) {
-            case Source.SourceStatus.CHARGEABLE:
-            case Source.SourceStatus.CONSUMED:
+            case Chargeable:
+            case Consumed:
               promise.resolve(convertSourceToWritableMap(source));
               break;
-            case Source.SourceStatus.CANCELED:
+            case Canceled:
               promise.reject(
-                      getErrorCode(mErrorCodes, "redirectCancelled"),
-                      getDescription(mErrorCodes, "redirectCancelled")
-              );
+                getErrorCode(mErrorCodes, "redirectCancelled"),
+                getDescription(mErrorCodes, "redirectCancelled"));
               break;
-            case Source.SourceStatus.PENDING:
-            case Source.SourceStatus.FAILED:
-            default:
+            case Failed:
+            case Pending:
               promise.reject(
-                      getErrorCode(mErrorCodes, "redirectFailed"),
-                      getDescription(mErrorCodes, "redirectFailed")
-              );
+                  getErrorCode(mErrorCodes, "redirectFailed"),
+                  getDescription(mErrorCodes, "redirectFailed"));
+              break;
           }
         }
       }
     }.execute();
   }
-
 }
