@@ -7,7 +7,9 @@
 //
 
 #import "TPSStripeManager.h"
-#import <Stripe/Stripe.h>
+#import <Stripe/Stripe-Swift.h>
+
+#import <stripe_payment/stripe_payment-umbrella.h>
 
 #import "TPSError.h"
 #import "TPSStripeManager+Constants.h"
@@ -263,6 +265,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
     RCTPromiseRejectBlock promiseRejector;
 
     BOOL requestIsCompleted;
+    BOOL applePayShouldCreatePaymentMethod;
 
     void (^applePayCompletion)(PKPaymentAuthorizationStatus);
     NSError *applePayStripeError;
@@ -301,12 +304,12 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
     publishableKey = options[@"publishableKey"];
     merchantId = options[@"merchantId"];
     errorCodes = errors;
-    [Stripe setDefaultPublishableKey:publishableKey];
+    [StripeAPI setDefaultPublishableKey:publishableKey];
 }
 
 -(void)setStripeAccount:(NSString *)_stripeAccount {
     NSString *_account;
-    if (_stripeAccount && _stripeAccount != [NSNull null]) {
+    if (_stripeAccount) {
         _account = _stripeAccount;
     }
     stripeAccount = _account;
@@ -332,7 +335,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
     NSArray <NSString *> *paymentNetworksStrings = [StripeModule applePaySupportedPaymentNetworksStrings];
 
     NSArray <PKPaymentNetwork> *networks = [self paymentNetworks:paymentNetworksStrings];
-    resolve([PKPaymentAuthorizationViewController canMakePaymentsUsingNetworks:networks] ? paymentNetworksStrings : nil);
+    resolve([PKPaymentAuthorizationViewController canMakePaymentsUsingNetworks:networks] ? paymentNetworksStrings : @[]);
 }
 
 
@@ -813,9 +816,28 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
 }
 
 -(void)paymentRequestWithApplePay:(NSArray *)items
-                  withOptions:(NSDictionary *)options
-                  resolver:(RCTPromiseResolveBlock)resolve
-                  rejecter:(RCTPromiseRejectBlock)reject {
+                      withOptions:(NSDictionary *)options
+                         resolver:(RCTPromiseResolveBlock)resolve
+                         rejecter:(RCTPromiseRejectBlock)reject {
+    applePayShouldCreatePaymentMethod = NO;
+    
+    [self startApplePay:items withOptions:options resolver:resolve rejecter:reject];
+}
+
+-(void)paymentMethodFromApplePay:(NSArray *)items
+                     withOptions:(NSDictionary *)options
+                        resolver:(RCTPromiseResolveBlock)resolve
+                        rejecter:(RCTPromiseRejectBlock)reject {
+    applePayShouldCreatePaymentMethod = YES;
+    
+    [self startApplePay:items withOptions:options resolver:resolve rejecter:reject];
+}
+
+-(void)startApplePay:(NSArray *)items
+    withOptions:(NSDictionary *)options
+    resolver:(RCTPromiseResolveBlock)resolve
+    rejecter:(RCTPromiseRejectBlock)reject {
+
     if(!requestIsCompleted) {
         NSDictionary *error = [errorCodes valueForKey:kErrorKeyBusy];
         reject(error[kErrorKeyCode], error[kErrorKeyDescription], nil);
@@ -855,7 +877,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
         [summaryItems addObject:summaryItem];
     }
 
-    PKPaymentRequest *paymentRequest = [Stripe paymentRequestWithMerchantIdentifier:merchantId country:countryCode currency:currencyCode];
+    PKPaymentRequest *paymentRequest = [StripeAPI paymentRequestWithMerchantIdentifier:merchantId country:countryCode currency:currencyCode];
 
     [paymentRequest setRequiredShippingAddressFields:requiredShippingAddressFields];
     [paymentRequest setRequiredBillingAddressFields:requiredBillingAddressFields];
@@ -1101,7 +1123,6 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
              TPSStripeParam(PaymentMethod, billingDetails): [self convertPaymentMethodBillingDetails: method.billingDetails] ?: NSNull.null,
              TPSStripeParam(PaymentMethod, card): [self convertPaymentMethodCard: method.card] ?: NSNull.null,
              TPSEntry(customerId),
-             TPSEntry(metadata),
              };
 #undef TPSEntry
 }
@@ -1182,7 +1203,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
 }
 
 - (BOOL)canSubmitPaymentRequest:(PKPaymentRequest *)paymentRequest rejecter:(RCTPromiseRejectBlock)reject {
-    if (![Stripe deviceSupportsApplePay]) {
+    if (![StripeAPI deviceSupportsApplePay]) {
         NSDictionary *error = [errorCodes valueForKey:kErrorKeyDeviceNotSupportsNativePay];
         reject(error[kErrorKeyCode], error[kErrorKeyDescription], nil);
         return NO;
@@ -1210,7 +1231,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
 
 - (void)addCardViewController:(STPAddCardViewController *)addCardViewController
        didCreatePaymentMethod:(STPPaymentMethod *)paymentMethod
-                   completion:(STPErrorBlock)completion {
+                   completion:(void (^)(NSError * _Nullable))completion {
     [RCTPresentedViewController() dismissViewControllerAnimated:YES completion:nil];
 
     requestIsCompleted = YES;
@@ -1237,27 +1258,52 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
     applePayCompletion = completion;
 
     STPAPIClient *stripeAPIClient = [self newAPIClient];
+    
+    if (applePayShouldCreatePaymentMethod) {
+        [stripeAPIClient createPaymentMethodWithPayment:payment completion:^(STPPaymentMethod * _Nullable paymentMethod, NSError * _Nullable error) {
+            self->requestIsCompleted = YES;
 
-    [stripeAPIClient createTokenWithPayment:payment completion:^(STPToken * _Nullable token, NSError * _Nullable error) {
-        self->requestIsCompleted = YES;
+            if (error) {
+                // Save for deffered use
+                self->applePayStripeError = error;
+                [self resolveApplePayCompletion:PKPaymentAuthorizationStatusFailure];
+            } else {
+                NSDictionary *result = [self convertPaymentMethod:paymentMethod];
+//                NSDictionary *extra = @{
+//                                        @"billingContact": [self contactDetails:payment.billingContact] ?: [NSNull null],
+//                                        @"shippingContact": [self contactDetails:payment.shippingContact] ?: [NSNull null],
+//                                        @"shippingMethod": [self shippingDetails:payment.shippingMethod] ?: [NSNull null]
+//                                        };
+//
+//                [result setValue:extra forKey:@"extra"];
 
-        if (error) {
-            // Save for deffered use
-            self->applePayStripeError = error;
-            [self resolveApplePayCompletion:PKPaymentAuthorizationStatusFailure];
-        } else {
-            NSDictionary *result = [self convertTokenObject:token];
-            NSDictionary *extra = @{
-                                    @"billingContact": [self contactDetails:payment.billingContact] ?: [NSNull null],
-                                    @"shippingContact": [self contactDetails:payment.shippingContact] ?: [NSNull null],
-                                    @"shippingMethod": [self shippingDetails:payment.shippingMethod] ?: [NSNull null]
-                                    };
+                [self resolvePromise:result];
+            }
 
-            [result setValue:extra forKey:@"extra"];
+        }];
 
-            [self resolvePromise:result];
-        }
-    }];
+    } else {
+        [stripeAPIClient createTokenWithPayment:payment completion:^(STPToken * _Nullable token, NSError * _Nullable error) {
+            self->requestIsCompleted = YES;
+
+            if (error) {
+                // Save for deffered use
+                self->applePayStripeError = error;
+                [self resolveApplePayCompletion:PKPaymentAuthorizationStatusFailure];
+            } else {
+                NSDictionary *result = [self convertTokenObject:token];
+                NSDictionary *extra = @{
+                                        @"billingContact": [self contactDetails:payment.billingContact] ?: [NSNull null],
+                                        @"shippingContact": [self contactDetails:payment.shippingContact] ?: [NSNull null],
+                                        @"shippingMethod": [self shippingDetails:payment.shippingMethod] ?: [NSNull null]
+                                        };
+
+                [result setValue:extra forKey:@"extra"];
+
+                [self resolvePromise:result];
+            }
+        }];
+    }
 }
 
 
@@ -1294,7 +1340,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
                                             url:TPSAppInfoURL];
     });
 
-    STPAPIClient * client = [[STPAPIClient alloc] initWithPublishableKey:[Stripe defaultPublishableKey]];
+    STPAPIClient * client = [[STPAPIClient alloc] initWithPublishableKey:[StripeAPI defaultPublishableKey]];
     client.appInfo = info;
     client.stripeAccount = stripeAccount;
 
@@ -1369,7 +1415,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
     [result setValue:@(source.livemode) forKey:@"livemode"];
     [result setValue:source.amount forKey:@"amount"];
     [result setValue:source.stripeID forKey:@"sourceId"];
-
+        
     // Flow
     [result setValue:[self sourceFlow:source.flow] forKey:@"flow"];
 
@@ -1487,7 +1533,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
             return @"discover";
         case STPCardBrandDinersClub:
             return @"diners";
-        case STPCardBrandMasterCard:
+        case STPCardBrandMastercard:
             return @"mastercard";
         case STPCardBrandUnknown:
         default:
@@ -1497,7 +1543,7 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
 
 /// API: https://stripe.com/docs/api/cards/object#card_object-brand
 - (NSString *)cardBrandAsPresentableBrandString:(STPCardBrand)inputBrand {
-    return STPStringFromCardBrand(inputBrand);
+    return [STPCardBrandUtilities stringFromCardBrand:inputBrand];
 }
 
 - (NSString *)cardFunding:(STPCardFundingType)inputFunding {
@@ -1525,6 +1571,23 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
         case STPSourceCard3DSecureStatusUnknown:
         default:
             return @"unknown";
+    }
+}
+
+- (NSString *)sourceStatus:(STPSourceStatus)status {
+    switch (status) {
+        case STPSourceStatusPending:
+            return @"pending";
+        case STPSourceStatusChargeable:
+            return @"chargeable";
+        case STPSourceStatusConsumed:
+            return @"consumed";
+        case STPSourceStatusCanceled:
+            return @"canceled";
+        case STPSourceStatusFailed:
+            return @"failed";
+        case STPSourceStatusUnknown:
+            return @"";
     }
 }
 
@@ -1572,31 +1635,13 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
     }
 }
 
-- (NSString *)sourceStatus:(STPSourceStatus)inputStatus {
-    switch (inputStatus) {
-        case STPSourceStatusPending:
-            return @"pending";
-        case STPSourceStatusChargeable:
-            return @"chargable";
-        case STPSourceStatusConsumed:
-            return @"consumed";
-        case STPSourceStatusCanceled:
-            return @"canceled";
-        case STPSourceStatusFailed:
-            return @"failed";
-        case STPSourceStatusUnknown:
-        default:
-            return @"unknown";
-    }
-}
-
 - (NSString *)sourceType:(STPSourceType)inputType {
     switch (inputType) {
         case STPSourceTypeBancontact:
             return @"bancontact";
         case STPSourceTypeGiropay:
             return @"giropay";
-        case STPSourceTypeIDEAL:
+        case STPSourceTypeiDEAL:
             return @"ideal";
         case STPSourceTypeSEPADebit:
             return @"sepaDebit";
@@ -1816,12 +1861,73 @@ void initializeTPSPaymentNetworksWithConditionalMappings() {
     dispatch_once(&onceToken, ^{
         NSMutableDictionary *mutableMap = [@{} mutableCopy];
 
+        
         if ((&PKPaymentNetworkAmex) != NULL) {
             mutableMap[TPSPaymentNetworkAmex] = PKPaymentNetworkAmex;
+        }
+        
+        if ((&PKPaymentNetworkChinaUnionPay) != NULL) {
+            mutableMap[TPSPaymentNetworkChinaUnionPay] = PKPaymentNetworkChinaUnionPay;
         }
 
         if ((&PKPaymentNetworkDiscover) != NULL) {
             mutableMap[TPSPaymentNetworkDiscover] = PKPaymentNetworkDiscover;
+        }
+
+        if (@available(iOS 12.0, *)) {
+            if ((&PKPaymentNetworkEftpos) != NULL) {
+                mutableMap[TPSPaymentNetworkEftpos] = PKPaymentNetworkEftpos;
+            }
+
+            if ((&PKPaymentNetworkElectron) != NULL) {
+                mutableMap[TPSPaymentNetworkElectron] = PKPaymentNetworkElectron;
+            }
+
+        }
+
+        if (@available(iOS 12.1.1, *)) {
+            if ((&PKPaymentNetworkElo) != NULL) {
+                mutableMap[TPSPaymentNetworkElo] = PKPaymentNetworkElo;
+            }
+            if ((&PKPaymentNetworkMada) != NULL) {
+                mutableMap[TPSPaymentNetworkIDCredit] = PKPaymentNetworkMada;
+            }
+        }
+
+        if ((&PKPaymentNetworkInterac) != NULL) {
+            mutableMap[TPSPaymentNetworkIDCredit] = PKPaymentNetworkInterac;
+        }
+        if (@available(iOS 10.1, *)) {
+            if ((&PKPaymentNetworkJCB) != NULL) {
+                mutableMap[TPSPaymentNetworkIDCredit] = PKPaymentNetworkJCB;
+            }
+        } else {
+            // Fallback on earlier versions
+        }
+        
+        if (@available(iOS 12.0, *)) {
+            if ((&PKPaymentNetworkMaestro) != NULL) {
+                mutableMap[TPSPaymentNetworkIDCredit] = PKPaymentNetworkMaestro;
+            }
+        } else {
+            // Fallback on earlier versions
+        }
+        if ((&PKPaymentNetworkPrivateLabel) != NULL) {
+            mutableMap[TPSPaymentNetworkPrivateLabel] = PKPaymentNetworkPrivateLabel;
+        }
+        if (@available(iOS 10.3, *)) {
+            if ((&PKPaymentNetworkQuicPay) != NULL) {
+                mutableMap[TPSPaymentNetworkQuicPay] = PKPaymentNetworkQuicPay;
+            }
+        } else {
+            // Fallback on earlier versions
+        }
+        if (@available(iOS 10.1, *)) {
+            if ((&PKPaymentNetworkSuica) != NULL) {
+                mutableMap[TPSPaymentNetworkQuicPay] = PKPaymentNetworkSuica;
+            }
+        } else {
+            // Fallback on earlier versions
         }
 
         if ((&PKPaymentNetworkMasterCard) != NULL) {
