@@ -1,4 +1,6 @@
-import 'package:flutter/foundation.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+part 'apple_pay_payment_request.g.dart';
 
 enum RequiredBillingAddressFields { all, name, email, phone, postal_address }
 
@@ -6,82 +8,58 @@ enum RequiredShippingAddressFields { all, name, email, phone, postal_address }
 
 enum ShippingType { shipping, delivery, store_pickup, service_pickup }
 
+@JsonSerializable(createFactory: false)
 class ApplePayPaymentOptions {
-  List<RequiredBillingAddressFields> requiredBillingAddressFields;
-  List<RequiredShippingAddressFields> requiredShippingAddressFields;
-  List<ShippingMethod> shippingMethod;
-  String currencyCode;
-  String countryCode;
-  ShippingType shippingType;
-  List<ApplePayItem> items;
+  final List<RequiredBillingAddressFields> requiredBillingAddressFields;
+  final List<RequiredShippingAddressFields> requiredShippingAddressFields;
+  final List<ShippingMethod> shippingMethod;
+  final String currencyCode;
+  final String countryCode;
+  final ShippingType shippingType;
+  final List<ApplePayItem> items;
 
-  ApplePayPaymentOptions(
-      {this.requiredBillingAddressFields,
-      this.requiredShippingAddressFields,
-      this.shippingMethod,
-      this.currencyCode,
-      this.countryCode,
-      this.shippingType,
-      this.items});
+  const ApplePayPaymentOptions({
+    this.requiredBillingAddressFields,
+    this.requiredShippingAddressFields,
+    this.shippingMethod,
+    this.currencyCode,
+    this.countryCode,
+    this.shippingType,
+    this.items,
+  });
 
-  Map<dynamic, dynamic> get json {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.requiredBillingAddressFields != null) {
-      data['requiredBillingAddressFields'] = this.requiredBillingAddressFields.map((b) => describeEnum(b)).toList();
-    }
-    if (this.requiredShippingAddressFields != null) {
-      data['requiredShippingAddressFields'] = this.requiredShippingAddressFields.map((s) => describeEnum(s)).toList();
-    }
-    if (this.shippingMethod != null) {
-      data['shippingMethod'] = this.shippingMethod.map((s) => s.toJson()).toList();
-    }
-    data['currencyCode'] = currencyCode;
-    if (this.shippingType != null) {
-      data['shippingType'] = shippingType;
-    }
-    data['countryCode'] = countryCode;
-    return data;
-  }
+  Map toJson() => _$ApplePayPaymentOptionsToJson(this);
 }
 
+@JsonSerializable()
 class ShippingMethod {
-  String amount;
-  String detail;
-  String id;
-  String label;
+  final String amount;
+  final String detail;
+  final String id;
+  final String label;
 
-  ShippingMethod({this.amount, this.detail, this.id, this.label});
+  const ShippingMethod({
+    this.amount,
+    this.detail,
+    this.id,
+    this.label,
+  });
 
-  factory ShippingMethod.fromJson(Map<String, dynamic> json) {
-    return ShippingMethod(
-      amount: json['amount'],
-      detail: json['detail'],
-      id: json['id'],
-      label: json['label'],
-    );
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.amount != null) data['amount'] = this.amount;
-    if (this.detail != null) data['detail'] = this.detail;
-    if (this.id != null) data['id'] = this.id;
-    if (this.label != null) data['label'] = this.label;
-    return data;
-  }
+  factory ShippingMethod.fromJson(Map json) => _$ShippingMethodFromJson(json);
+  Map toJson() => _$ShippingMethodToJson(this);
 }
 
+@JsonSerializable(createFactory: false)
 class ApplePayItem {
-  String label;
-  String amount;
-  String type;
-  ApplePayItem({this.label, this.amount, this.type});
+  final String label;
+  final String amount;
+  final String type;
 
-  Map<String, dynamic> get json {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.label != null) data['label'] = this.label;
-    if (this.amount != null) data['amount'] = this.amount;
-    if (this.type != null) data['type'] = this.type;
-    return data;
-  }
+  const ApplePayItem({
+    this.label,
+    this.amount,
+    this.type,
+  });
+
+  Map toJson() => _$ApplePayItemToJson(this);
 }
